changes made for UX
2. TEXT
For the text about how to use PepVEP, I recommend using the active voice
which is perceived as more concise, clearer and inviting to the user.

For example, you could change this sentence: "Variants can be submitted
via pasting in the box in VCF or HGVS format, uploading a file in VCF
format or using the PepVEP API"
to -> "Submit variants by pasting them in the box below, uploading a
file or by using the PepVEP API"

One of the general rules of thumb for how people read on the web is that
they avoid reading word-by-word, they scan. So to help them quickly
understand content and also make them more likely to actually look at
chunks of text, it's good to simplify as much as possible.

For example, you could replace this sentence: "Further help and
explanations about the data in PepVEP can be found in the "about"
section at the top right hand side of the page."
with -> "More help and information is available in the About section."
(make About link or a trigger that opens About panel)

"We continually strive to make PepVEP clear and useful to our users, to
contact PepVEP with questions or suggestions please use the "contact"
link at the top of the page."
-> "We strive to make PepVEP clear and useful to our users. Contact us
with questions or suggestions here." (make here link or trigger Contact
panel)

3. BOXES

For the input boxes, I would recommend using the same voice and one that
is calling to action instead of it serving as a title for the box:

    * "Paste variants" - good
    * "File Upload" could become "Upload files"
    * "PepVEP REST API" could become "Access PepVEP REST API"

Regarding the box UI, I suggest a few quick fixes:

    * Add more padding, especially for the API box where the last text line
is too close to the button
    * Enlarge paste variants box to show the average input without scroll.
At the moment, if you populate the box with the examples, you need to
scroll to see all
    * Change input text colour to black - grey can be difficult to read
when set against white background because of low contrast, and it also
gives the impression of disabled text
    * Make call to action clear: make buttons with spacing on all sides
instead of the current spanning across the box width, so that it is
clear to the user that these are clickable
    * Make call to action text (Run, Upload file, PepVEP REST API) bold and
darker in colour to increase contrast between the button fill colour and
text

Instead of the three box approach, we could explore different design
versions where the user could preselect a way to input variants.

4. MISCELLANEOUS NOTES

    * I would suggest that at a later stage you implement Contact form
instead of sharing the helpdesk email address. Unfortunately, email
addresses are easily picked up and we receive lots of SPAM through
UniProt helpdesk email.
    * For consistency, put hgvs to all caps in this sentence in Upload list
box: "PepVEP also supports hgvs in below format"
    * Bug report: Can't reach the end of the About box

5. RESULTS TABLE

The results table looks good to me! Potential improvements could be:

    * a legend for CADD colour indications
    * tooltips for Annotations columns illustrations